move Supplier query results to related namespace
